Change advice on reuse of meta object
Since due to #155 which changes the behavior on content-length
recalculation the issue arises that if the meta object is send back as
is the content-length will be wrong there should be documentation for
this behavior, since also the reverting the patch causes other issues.

After all reusing all the headers as is might cause trouble anyway see
ISSUE #185, I would suggest changing the advice to only reuse the parts
of meta which explicitly are 'right'.